requirements.txt
-f /usr/share/pip-wheels
asn1crypto==0.24.0
beautifulsoup4==4.6.3
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
dj-database-url==0.5.0
Django==1.11.16
django-addendum==0.5.2
django-app-namespace-template-loader==0.4.1
django-appconf==1.0.2
django-autotranslate==1.1.0
django-blog-zinnia==0.20
django-ckeditor==5.6.1
django-constance==2.3.1
django-contrib-comments==1.9.0
django-crispy-forms==1.7.2
django-cryptography==0.3
django-fluent-comments==2.1
django-heroku==0.3.1
django-htmlmin==0.10.0
django-jet==1.0.8
django-js-asset==1.1.0
django-mptt==0.9.1
django-otto-admin==0.1.2
django-picklefield==1.1.0
django-pipeline==1.6.14
django-robots==3.1.0
django-rq==1.2.0
django-storages==1.7.1
django-suit==0.2.26
django-tag-parser==3.1
django-tagging==0.4.6
django-threadedcomments==1.2
django-widget-tweaks==1.4.3
django-xmlrpc==0.1.8
djangorestframework==3.9.0
docutils==0.14
futures==3.1.1
google-api-core==1.5.2
google-api-python-client==1.4.1
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
goslate==1.5.1
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.12.0
idna==2.7
manifesto==0.5
mots-vides==2015.5.11
mysqlclient==1.3.13
oauth2client==1.5.2
Pillow==5.3.0
polib==1.1.0
protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pyparsing==2.3.0
python-akismet==0.4.1
pytz==2018.5
redis==3.0.1
regex==2018.11.7
requests==2.20.1
rq==0.12.0
rsa==4.0
six==1.11.0
sorl-thumbnail==12.5.0
uritemplate==3.0.0
urllib3==1.24.1
webencodings==0.5.1
whitenoise==4.1.2
zinnia-wysiwyg-ckeditor==1.3
